app/requirements.txt
annotated-types==0.7.0
anyio==4.3.0
click==8.1.8
fastapi==0.110.1
h11==0.16.0
idna==3.7
Jinja2==3.1.3
joblib==1.3.2
MarkupSafe==2.1.5
numpy==1.26.4
pandas==2.2.2
pydantic==2.6.4
python-dateutil==2.9.0.post0
pytz==2024.1
scikit-learn==1.4.2
scipy==1.13.1
six==1.16.0
sniffio==1.3.0
starlette==0.37.2
threadpoolctl==3.5.0
typing-inspection==0.4.0
typing_extensions>=4.12.0
tzdata==2024.1
uvicorn==0.29.0
huggingface_hub==0.23.0
python-multipart==0.0.9
python-dotenv==1.0.1
